[Android] Schedule the XWalkView/Webview's API called on UI thread.
If taken XWalkView as backend, it requires all of its API should
be called from UI thread. But the inappbrowser plugin breaks this rule,
for example, it would call the getUrl()/loadUrl() API of XWalkView. This
would cause XWalkView throw RuntimeException.

So refine the inappbrowser plugin, and schedule the webview's API called
on the UI thread. Btw, this plugin has the same issue if taken
Android WebView as backend after the API lever >=18.

BUG=https://github.com/otcshare/cordova-xwalk-android/issues/70